pyproject.toml
[tool.poetry]
name = "fastapi-ca"
version = "0.1.0"
description = ""
authors = ["Jungkihong07 <hnn07185@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
py-ulid = "^1.0.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.36"
mysqlclient = "^2.2.6"
alembic = "^1.14.0"
dependency-injector = "^4.45.0"
black = "^25.1.0"
pydantic = {extras = ["email"], version = "^2.10.6"}
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.8.0"
celery = "^5.4.0"
eventlet = "^0.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
